pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "joblib>=1.5.2",
    "phonenumbers>=9.0.15",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "scikit-learn>=1.7.2",
    "uvicorn>=0.37.0",
    "validators>=0.35.0",
]
